backend/pyproject.toml
[project]
name = "beautyback"
version = "0.1.0"
description = "BeautyLab Backend API - FastAPI with PostgreSQL, JWT Auth, and Dynamic Theming"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic==1.18.4",
    "annotated-doc==0.0.4",
    "annotated-types==0.7.0",
    "anyio==4.12.1",
    "argon2-cffi==25.1.0",
    "argon2-cffi-bindings==25.1.0",
    "asyncpg==0.31.0",
    "certifi==2026.2.25",
    "cffi==2.0.0",
    "click==8.3.1",
    "dnspython==2.8.0",
    "ecdsa==0.19.1",
    "email-validator==2.3.0",
    "fastapi==0.135.1",
    "fastapi-cli==0.0.24",
    "fastapi-cloud-cli==0.14.0",
    "fastar==0.8.0",
    "greenlet==3.3.2",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httptools==0.7.1",
    "httpx==0.28.1",
    "idna==3.11",
    "jinja2==3.1.6",
    "mako==1.3.10",
    "markdown-it-py==4.0.0",
    "markupsafe==3.0.3",
    "mdurl==0.1.2",
    "pyasn1==0.6.2",
    "pycparser==3.0",
    "pydantic==2.12.5",
    "pydantic-core==2.41.5",
    "pydantic-extra-types==2.11.0",
    "pydantic-settings==2.13.1",
    "pygments==2.19.2",
    "python-dotenv==1.2.2",
    "python-jose==3.5.0",
    "python-multipart==0.0.22",
    "pyyaml==6.0.3",
    "rich==14.3.3",
    "rich-toolkit==0.19.7",
    "rignore==0.7.6",
    "rsa==4.9.1",
    "ruff==0.15.4",
    "sentry-sdk==2.54.0",
    "shellingham==1.5.4",
    "six==1.17.0",
    "sqlalchemy==2.0.48",
    "starlette==0.52.1",
    "typer==0.24.1",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "urllib3==2.6.3",
    "uvicorn==0.41.0",
    "uvloop==0.22.1",
    "watchfiles==1.1.1",
    "websockets==16.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
